Do floating point calculations in ... floating point.  (Bugfix)
git-svn-id: file:///home/xapian-svn/xapian/trunk@221 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/backends/multi/multi_database.h
@@ -133,7 +133,8 @@ inline MultiTermList::MultiTermList(TermList *tl_new,
 				    const IRDatabase *rootdb_new)
 	: tl(tl_new), termdb(termdb_new), rootdb(rootdb_new)
 {
-    termfreq_factor = (rootdb->get_doccount()) / (termdb->get_doccount());
+    termfreq_factor = ((double)(rootdb->get_doccount())) /
+		      (termdb->get_doccount());
 printf("Approximation factor for termfrequency: %f\n", termfreq_factor);
 }
